addUpdateMerch

+++ app/Http/Controllers/MerchController.php
@@ -42,16 +42,46 @@ class MerchController extends Controller
      }
      public function delete($id)
      {
-
-        $merch = Merch::find($id);
-      //  dd('/storage'.'/'.$merch->nameMainPhoto);
-        unlink(public_path('/storage'.'/'.$merch->nameMainPhoto));
-        unlink(public_path('/storage'.'/'.$merch->namePhoto1));
-        unlink(public_path('/storage'.'/'.$merch->namePhoto2));
-
-       Merch::find($id)->delete();
-       //Merch::find($id)->delete();
+      $merch = Merch::find($id);
+        if($merch)
+        { 
+              unlink(public_path('/storage//'.$merch->nameMainPhoto));
+              unlink(public_path('/storage//'.$merch->namePhoto1));
+              unlink(public_path('/storage//'.$merch->namePhoto2));
+              Merch::find($id)->delete();  
+        }
        return view('merch\showMerch',['Merch'=>Merch::all()]);
+     }
+     public function updateMerch($id)
+     {
+      return view('merch\updateMerch',['Merch'=>Merch::find($id)]);
+     }
+     public function update($id, Request $request)
+     {
+       $merch = Merch::find($id);
+
+       if($request->file('fileMain'))
+       {
+        $photoMainName = $request->file('fileMain')->store('uploads','public');
+        $merch->nameMainPhoto = $photoMainName;
+       }
+       if($request->file('file1'))
+       {
+        $photoName1 = $request->file('file1')->store('uploads','public');
+        $merch->namePhoto1= $photoName1;
+       }
+
+       if($request->file('file2'))
+       {
+        $photoName2 = $request->file('file2')->store('uploads','public');
+        $merch->namePhoto2= $photoName2;
+       }
+       
+       $merch->title= $request->input('title');
+       $merch->describe= $request->input('describe');
+       $merch->save();
+
+      return view('merch\merchOne',['data'=>Merch::find($id)]);
      }
 
 }
